api controller and correct data format

+++ CloudOnWebApp/Entities/Product.cs
@@ -9,11 +9,11 @@ namespace CloudOnWebApp.Entities
     {
         public int Id { get; set; }
 
-        public int ExternalId { get; set; }
-        public int Code { get; set; }
+        public long ExternalId { get; set; }
+        public long Code { get; set; }
         public string Description { get; set; }
         public string Name { get; set; }
-        public int Barcode { get; set; }
+        public long Barcode { get; set; }
         public double RetailPrice { get; set; }
         public double WholePrice { get; set; }
         public double Discount { get; set; }
